added sweet alerts

+++ server/public/scripts/client.js
@@ -3,7 +3,9 @@ console.log('initializing client...')
 //Gobal Variables
 //tasks = [];     //array to hold data on the client side
 let complete = '';  //changes true false to yes no.
+let completeClass = ''; //Changes button class/appearance if completed or not.
 let element;
+
 
 //initialize jquery
 $(document).ready(onReady);
@@ -14,9 +16,11 @@ function onReady () {
     //event listeners
     $(document).on("click", "#addbtn", addTask);
     $(document).on("click", "#deletebtn", deleteTask);
+    $(document).on("click", "#completebtn", updateComplete);
     $(document).on("click", ".pikachu", scrollUp);
     $(document).on("click", ".squidward", scrollDown);
     $(document).on("click", ".unicorn", scrollDown);
+
 
 
     //get dynamic data
@@ -40,6 +44,25 @@ function scrollDown() {
 //clear input from text box
 function clearInput() {
     $('#name').attr('placeholder', 'Enter');
+}
+
+
+//function to update complete
+function updateComplete () {
+    console.log('in UpdateComplete');
+    let id = $(this).parents("tr").data("id");
+
+    $.ajax({
+        method: "PUT",
+        url: `/tasks/${id}`,
+        data: { complete: true },
+      })
+      .then(() => {
+        getTasks();
+      })
+      .catch((err) => {
+        console.log("PUT failed", err);
+      });
 }
 
 
@@ -71,19 +94,37 @@ function addTask() {
 function deleteTask() {
     console.log('In deleteTask');
 
-    let id = $(this).parent('tr').data('id');
+    //sweet alert
+    swal({
+        title: "Warning!",
+        icon: "warning",
+        text: "Delete Task?",
+        buttons: true,
+        dangerMode: true,
+    })
 
-    $.ajax ({
-        method: 'DELETE',
-        url: `/tasks/${id}`
-    })
-    .then (() => {
-        getTasks();
+    .then ((willDelete) => {
 
-    })
-    .catch ((error) => {
-        console.log('ERROR could not DELETE', error);
-    })
+        if(willDelete) {
+            let id = $(this).parent('tr').data('id');
+
+            $.ajax ({
+                method: 'DELETE',
+                url: `/tasks/${id}`
+            })
+            .then (() => {
+                getTasks();
+
+            })
+            .catch ((error) => {
+                console.log('ERROR could not DELETE', error);
+            })
+        }
+        else {
+            swal("Task saved!");
+        }
+
+    });
 
 }
 
@@ -117,18 +158,22 @@ function render(tasks) {
         console.log(task);
 
         if (task.complete) {
-            complete = 'Yes';
+            complete = '<img width="20%" height="auto" src="./views/greencircle.png">';
+            completeClass = 'taskComplete'
+            console.log('completeClass is', completeClass);
           }
           else {
-            complete = 'No';
+            complete = '<img class="squidward" src="./views/goldexlamation.png">';
+            completeClass = 'taskPending';
+            console.log('completeClass is', completeClass);
           }
 
         $('#viewTasks').append(`
-            <tr class="taskWrapper" data-id=${task.id} data-complete=${complete}>
+            <tr class="taskWrapper" data-id=${task.id} data-complete="${complete}">
                 <td id="deletebtn" class="taskList">
                     ${task.name}
                 </td>
-                <td id="completebtn" class="taskList">
+                <td id="completebtn" class="taskList ${completeClass}">
                     ${complete}
                 </td> 
             </tr>
